Move original test into its own file so we can build on it with many more tests
And the custom metadata attribute tests class wasn't appropriate because the issue that was being tested was metadata view handling -- the fact that the metadata view class was also a custom metadata attribute is purely incidental to VS MEF.
